Lesson1

+++ src/main/java/ru/homework/java/core/Race.java
@@ -2,5 +2,5 @@ package ru.homework.java.core;
 
 public interface Race {
     boolean run(String name, int lengthLimit, int distance);
-    boolean jump(String name, int heightLimit, int height);
+    boolean jump(String name, double heightLimit, double height);
 }
